replace UsagePlanKey with UsagPlan.addApiKey
