+++ packages/@aws-cdk/aws-apigateway/lib/index.ts
@@ -9,7 +9,6 @@ export * from './integrations';
 export * from './lambda-api';
 export * from './api-key';
 export * from './usage-plan';
-export * from './usage-plan-key';
 
 // AWS::ApiGateway CloudFormation Resources:
 export * from './apigateway.generated';
