+++ packages/@aws-cdk/aws-apigateway/test/integ.restapi.ts
@@ -85,10 +85,7 @@ class Test extends cdk.Stack {
       resources: [api.root]
     });
 
-    new apigateway.UsagePlanKey(this, 'UsagePlanKey', {
-      apiKey,
-      usagePlan
-    });
+    usagePlan.addApiKey(apiKey);
   }
 }
